Merged PR 432444: Preaggregation: Handle Count Aggregate Expression when data has null and LA is pushed only on non aggregate side of join
Preaggregation doesn't give correct result when data has null and aggregate has count as aggregate function and LA is pushed only on non aggregate side of join.

Count(y) Group by x gives sum of rows for each x where y is not null but after preaggregation this used to give sum of rows irrespective of if it is null or not. This PR fixes this.

Related work items: #897999