Add FetchBusyChecker
The FetchBusyChecker class is responsible for monitoring the fetch latency of the system and determining whether it is in a fetch busy mode.
It keeps track of the fetch latency and switches the system to fetch busy mode when the latency exceeds a certain threshold (highWatermarkMs).
Once in fetch busy mode, the system will remain in that state until the fetch latency falls below another threshold (lowWatermarkMs)
and a cooldown period (coolDownMs) has passed since the last fetch busy time.

b/313438223

Change-Id: I9419ad1f7fa80cf9bcbc6c0d0d3450fbf73495e2
Reviewed-on: https://bigdataoss-internal-review.googlesource.com/c/third_party/apache/spark/+/73740
Tested-by: Prow Service Account <425329972751-compute@developer.gserviceaccount.com>
Reviewed-by: Igor Dvorzhak <idv@google.com>